Split healthcheck-related stuff out of domain/common.go into new domain/health.go, and Merge isvcs.HealthCheckStatus and health.healthStatus into domain.HealthCheckStatus
